feat (clrs.ch24.s23): implement and test the dag algorithm for finding shortest paths

+++ cs-clrs/src/main/java/com/mmnaseri/cs/clrs/ch24/s2/BellmanFordSingleSourceShortestPathFinder.java
@@ -31,7 +31,7 @@ public class BellmanFordSingleSourceShortestPathFinder<E extends WeightedEdgeDet
             final Vertex<V> midpoint = edge.getFrom();
             final Vertex<V> destination = edge.getTo();
             final int currentDistance = destination.getProperty("distance", Integer.class);
-            final int detourDistance = midpoint.getProperty("distance", Integer.class) + weight(edge);
+            final long detourDistance = (long) midpoint.getProperty("distance", Integer.class) + (long) weight(edge);
             if (currentDistance > detourDistance) {
                 return null;
             }
